Split Handle from RPCClient.h for generate swig
