+++ bcos-sdk/SWIG/TarsClient.i
@@ -2,5 +2,11 @@
 #include "bcos-cpp-sdk/tarsRPC/RPCClient.h"
 %}
 
-%include <std_vector.i>
+%include <std_shared_ptr.i>
+
+%include "../bcos-cpp-sdk/tarsRPC/Handle.h"
+%template(HandleTransactionReceipt) bcos::sdk::Handle<bcos::protocol::TransactionReceipt::Ptr>;
+%template(HandleLong) bcos::sdk::Handle<long>;
+%shared_ptr(bcos::sdk::Callback)
+
 %include "../bcos-cpp-sdk/tarsRPC/RPCClient.h"
