+++ bcos-sdk/bcos-cpp-sdk/tarsRPC/Handle.h
@@ -1,0 +1,68 @@
+#pragma once
+#include "detail/Core.h"
+#include <servant/Communicator.h>
+#include <future>
+
+namespace bcos::sdk
+{
+class RPCClient;
+
+class Callback
+{
+public:
+    Callback() = default;
+    Callback(const Callback&) = default;
+    Callback(Callback&&) noexcept = default;
+    Callback& operator=(const Callback&) = default;
+    Callback& operator=(Callback&&) noexcept = default;
+
+    virtual ~Callback() noexcept = default;
+    virtual void onMessage() = 0;
+};
+
+template <class Response>
+class Handle
+{
+private:
+    RPCClient& m_rpcClient;
+    std::future<tars::ReqMessagePtr> m_future;
+    std::shared_ptr<Callback> m_callback;
+
+protected:
+    RPCClient& rpcClient() { return m_rpcClient; }
+    void setFuture(std::future<tars::ReqMessagePtr> future) { m_future = std::move(future); }
+    std::shared_ptr<Callback> callback() { return m_callback; }
+
+public:
+    Handle(RPCClient& rpcClient) : m_rpcClient(rpcClient){};
+    Handle(Handle const&) = delete;
+    Handle& operator=(Handle const&) = delete;
+    Handle(Handle&&) noexcept = default;
+    Handle& operator=(Handle&&) noexcept = default;
+    ~Handle() noexcept = default;
+
+    Response get()
+    {
+        auto message = m_future.get();
+        message->callback->dispatch(message);
+
+        auto& tarsCallback = dynamic_cast<detail::TarsCallback&>(*message->callback);
+        return std::move(tarsCallback).takeResponse<Response>();
+    }
+    void waitFor() { m_future.wait(); }
+    template <typename Rep, typename Period>
+    std::future_status waitFor(const std::chrono::duration<Rep, Period>& rel)
+    {
+        return m_future.wait_for(rel);
+    }
+    template <typename Clock, typename Duration>
+    std::future_status waitUntil(const std::chrono::time_point<Clock, Duration>& abs)
+    {
+        return m_future.wait_until(abs);
+    }
+    void setCallback(std::shared_ptr<bcos::sdk::Callback> callback)
+    {
+        m_callback = std::move(callback);
+    }
+};
+}  // namespace bcos::sdk
